Refactor: move dashboard components to a single file
