+++ src/components/Dashboard/AQLegend.tsx
