+++ src/components/Dashboard/AirQualityGroupDeviceDataLogic.tsx
