+++ src/pages/dashboard/air-quality.tsx
@@ -4,12 +4,12 @@ import { Typography, Tooltip, makeStyles } from "@material-ui/core";
 import { Marker } from "react-map-gl";
 import WithLoading from "components/WithLoading";
 import Meta from "components/Meta";
-import DashboardLayout from "components/DashboardLayout";
+import DashboardLayout from "components/Dashboard/DashboardLayout";
 import { fetcher } from "utils";
 import { MapPinIcon } from "../../constants";
 import Map from "components/Dashboard/Map";
 import { Device } from "lib/aqmd";
-import AirQualityGroupDeviceDataLogic from "components/AirQualityGroupDeviceDataLogic";
+import AirQualityGroupDeviceDataLogic from "components/Dashboard/AirQualityGroupDeviceDataLogic";
 import PurpleAirSensorData from "purple-air-data.json";
 import AeroqualSensor from "aeroqual-sensor.json";
 import { AirQualityDevices } from "types";
