+++ src/pages/dashboard/water-quality.tsx
@@ -13,10 +13,10 @@ import { groupBy } from "lodash";
 import { Marker } from "react-map-gl";
 import { Parameter, Data, SiteData, DashboardPage } from "types";
 import { colorScale, getAverage, getColorFromScale, getRange } from "utils";
-import DashboardLayout from "components/DashboardLayout";
+import DashboardLayout from "components/Dashboard/DashboardLayout";
 import Map from "components/Dashboard/Map";
 import Table from "components/Dashboard/Table";
-import DownloadDataButtonsSection from "components/DownloadDataButtonsSection";
+import DownloadDataButtonsSection from "components/Dashboard/DownloadDataButtonsSection";
 import WithLoading from "components/WithLoading";
 import Meta from "components/Meta";
 import ContinuousColorLegend from "components/ContinuousColorLegend";
